move stridx to server directory
