+++ server/server.lua
@@ -4,7 +4,7 @@ local socket = require("socket")
 local logger = require("logger")
 require("class")
 json = require("dkjson")
-require("stridx")
+require("server.stridx")
 require("gen_panels")
 require("csprng")
 local NetworkProtocol = require("NetworkProtocol")
