+++ server/stridx.lua
